+++ opcl.h
@@ -1,3 +1,3 @@
 #include "../opencl10/CL/opencl.h"
 
-void opencl_init(char*, int **, int);
+void opencl_init(char*, float **, int);
